+++ apps/docs/stories/components/BankAccountForm/example.ts
@@ -61,7 +61,7 @@ ${codeExampleHead(
       console.error(event.detail);
     });
 
-    document.addEventListener('checkout-complete-event', (event) => {
+    document.addEventListener('submit-event', (event) => {
       console.log('Checkout completed successfully!', event.detail);
     });
   })();