+++ apps/docs/stories/components/SezzlePaymentMethod/example.ts
@@ -65,7 +65,7 @@ ${codeExampleHead(
       console.error(event.detail);
     });
 
-    document.addEventListener('checkout-complete-event', (event) => {
+    document.addEventListener('submit-event', (event) => {
       console.log('Checkout completed successfully!', event.detail);
     });
   })();