+++ packages/webcomponents/src/components/insurance/season-interruption/season-interruption-insurance.tsx
@@ -5,11 +5,10 @@ import { ComponentErrorCodes, ComponentErrorSeverity } from '../../../api/Compon
 import JustifiAnalytics from '../../../api/Analytics';
 import { checkPkgVersion } from '../../../utils/check-pkg-version';
 import { ComponentErrorEvent } from '../../../api/ComponentEvents';
-import checkoutStore from '../../../store/checkout.store';
+import { checkoutStore } from '../../../store/checkout.store';
 
 @Component({
   tag: 'justifi-season-interruption-insurance',
-  shadow: true,
 })
 export class SeasonInterruptionInsurance {
   @Prop() authToken?: string;